requirements.txt
streamlit 
pandas 
pymongo 
faiss-cpu 
sentence-transformers

